Seperate Meld, Vanish and Shade
Having added the Meld visual effects to Vanish to explain the concealed
move, it was becoming confusing what was Meld and what wasn't.

Hence the Meld visuals are now part of an effect called Shade. Meld
induces Shade, as does Vanish.

(We don't use Shade in the UI; instead the flyover text popup says
either "Vanish" or "Meld" as appropriate. (It picks up the "Meld" name
since we've now added a localized friendly name for Spook_MeldTrigger).

Like Chryssalid burrow, Shade also now triggers a visiblity
recalculation when the effect ends, to allow enemies to detect a spook
who walks up to them under Vanish and just stands next to them in open
ground, once the Vanish (or more specifically, the Shade effect it
produced) wears off, which is pretty immediately. This gives enemies the
chance to immediately slaughter the offending spook.

Added comments (X2Effect_SpookShade) to the effect that shade is not a
separate kind of gameplay invisiblity. We could make it one through
further extensions to the tactical detection manager, but have chosen
not to. Instead shade is just one kind of concealment preservation, and
it's concealment that provides gameplay invisibility. Standing in a
location from which a spook can Meld is another kind of concealment
preservation. (Triggering Meld to induce Shade isn't required for
concealment preservation; the location is safe regardless of the
presence or absence of the Shade effect, so that the Spook doesn't
immediately get revealed by nearby enemies at the start of their next
turn after a Meld-induced-Shade wears off. In this sense the
Meld-induced Shade is partly cosmetic; we say that Shade provides
concealment preservation so that Vanish causes such, but we don't solely
use Shade as the determiner of concealment preservation.

If we changed the Meld-induced Shade to not wear off until the unit
moved, we could drop the "Meldable location" part of our concealment
preservation and just use Shade. That might make more sense, from
a purist perspective, and would be visually more consistent too.