Fix Eclipse mind-control sensitivity
Eclipse previously worked on hostile units then mind-controlled to be
on XCOM's team, which it shouldn't have.

Eclipse previously didn't work on friendly units then mind-controlled to
on the alien team, which it should have.

This commit uses our custom condition to detect the original teams of
both units and check they are the same, i.e. the units were originally
allies.